[refactor] AllergyType, AllergyProperties
1. entity 의 의미가 데이터베이스 모델링에 국한되지 않는다는 판단 하에 entity packaging
2. AllergyType.nameOf 로직 개선
3. AllergyProperties 에 getAllergyByIngredient, isAllergenicIngredient 메소드 추가
