+++ src/main/java/allercheck/backend/domain/allergy/entity/AllergyProperties.java
@@ -1,4 +1,4 @@
-package allercheck.backend.domain.allergy.application;
+package allercheck.backend.domain.allergy.entity;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -17,5 +17,13 @@ public class AllergyProperties {
     private final Map<String, AllergyType> allergenicIngredients;
     public AllergyProperties(@Value("#{${allergenicIngredientsMap}}") Map<String, AllergyType> allergenicIngredientsMap) {
         this.allergenicIngredients = Map.copyOf(allergenicIngredientsMap);
+    }
+
+    public AllergyType getAllergyByIngredient(String ingredient) {
+        return allergenicIngredients.get(ingredient);
+    }
+
+    public boolean isAllergenicIngredient(String ingredient) {
+        return allergenicIngredients.containsKey(ingredient);
     }
 }
