+++ src/main/java/allercheck/backend/domain/allergy/exception/InvalidAllergyTypeNameException.java
@@ -1,0 +1,4 @@
+package allercheck.backend.domain.allergy.exception;
+
+public class InvalidAllergyTypeNameException extends RuntimeException{
+}
